requirements.txt
asgiref==3.8.1
Django==5.2.3
gunicorn==21.2.0
pip==25.1.1
python-dotenv==1.1.1
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
wheel==0.45.1

# Database
psycopg2-binary==2.9.10

# Django Rest Framework
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0

# Tailwind CSS
django-tailwind==3.8.0
django-browser-reload==1.12.1

# Testing dependencies
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0

# Code quality
black==24.1.1
flake8==7.0.0 